+++ CHANGELOG.md
@@ -8,6 +8,7 @@ x.x.x Release notes (yyyy-MM-dd)
 ### Enhancements
 
 * `-[RLMRealm inWriteTransaction]` is now public.
+* Exceptions raised when incorrect object types are used with predicates now contain more detailed information.
 
 ### Bugfixes